Fix to group claimants page
