+++ features/support/capybara_selectors/group_claimants.rb
@@ -1,6 +1,7 @@
 Capybara.add_selector(:group_claimants) do
-  xpath do |locator, _options|
-    translated = EtFullSystem::Test::Messaging.instance.translate(locator)
-    XPath.generate { |x| x.descendant(:fieldset)[x.descendant(:legend)[x.string.n.is("translated #{number}")]] }
+  xpath do |locator, number:, **_options|
+    translated = EtFullSystem::Test::Messaging.instance.translate(locator, number: number)
+    XPath.generate { |x| x.descendant(:fieldset)[x.child(:legend)[x.string.n.is(translated)]] }
   end
+  expression_filter(:number) { |xpath, _val| xpath }
 end
