+++ test_common/page_objects/et1/group_claimants_page.rb
@@ -43,74 +43,13 @@ module EtFullSystem
             element :six_more_claimants, :panel_indent, 'claims.additional_claimants.csv_upload_text_html', exact: false
             element :csv_upload_link, :link_named, 'claims.additional_claimants.csv_upload_link', exact: false
           end
-            (2..6).each_with_index do |number, idx|
-            section :"about_claimant_#{number}", :xpath, (XPath.generate { |x| x.descendant(:fieldset)[x.descendant(:legend)[x.string.n.is("Claimant #{idx + 2}")]] }) do
-              # section :"about_claimant_#{number}", :group_claimants, "claims.additional_claimants.person", number: idx+2 do
-                section :title, :question_labelled, 'simple_form.labels.claimant.title' do
-                  include ::EtFullSystem::Test::I18n
-                  def set(value)
-                    root_element.select(factory_translate(value))
-                  end
-                end
-                #first name
-                section :first_name, :question_labelled, 'simple_form.labels.claimant.first_name' do
-                  element :field, :css, 'input'
-                  delegate :set, to: :field
-                end
-                #lastname name
-                section :last_name, :question_labelled, 'simple_form.labels.claimant.last_name' do
-                  element :field, :css, 'input'
-                  delegate :set, to: :field
-                end
-                #Date of birth
-                section :date_of_birth, :legend_header, 'claims.personal_details.date_of_birth', exact: false do
-                  element :date_of_birth_hint, :paragraph, 'simple_form.hints.claimant.date_of_birth'
-                  section :day, :question_labelled, 'simple_form.labels.claimant.date_of_birth.day' do
-                    element :field, :css, 'input'
-                    delegate :set, to: :field
-                  end
-                  section :month, :question_labelled, 'simple_form.labels.claimant.date_of_birth.month' do
-                    element :field, :css, 'input'
-                    delegate :set, to: :field
-                  end
-                  section :year, :question_labelled, 'simple_form.labels.claimant.date_of_birth.year' do
-                    element :field, :css, 'input'
-                    delegate :set, to: :field
-                  end
-                  def set(value)
-                    (day_value, month_value, year_value) = value.split("/")
-                    day.set(day_value)
-                    month.set(month_value)
-                    year.set(year_value)
-                  end
-                end
-                #Building number or name
-                section :building, :question_labelled, 'simple_form.labels.claimant.address_building' do
-                  element :field, :css, 'input'
-                  delegate :set, to: :field
-                end
-                #Street
-                section :street, :question_labelled, 'simple_form.labels.claimant.address_street' do
-                  element :field, :css, 'input'
-                  delegate :set, to: :field
-                end
-                #Town/city
-                section :locality, :question_labelled, 'simple_form.labels.claimant.address_locality' do
-                  element :field, :css, 'input'
-                  delegate :set, to: :field
-                end
-                #County
-                section :county, :question_labelled, 'simple_form.labels.claimant.address_county' do
-                  element :field, :css, 'input'
-                  delegate :set, to: :field
-                end
-                #Postcode
-                section :post_code, :question_labelled, 'simple_form.labels.claimant.address_post_code' do
-                  element :field, :css, 'input'
-                  delegate :set, to: :field
-                end
-              end
-            end
+
+          section :about_claimant_2, GroupClaimantSection, :group_claimants, "claims.additional_claimants.person", number: 2
+          section :about_claimant_3, GroupClaimantSection, :group_claimants, "claims.additional_claimants.person", number: 3
+          section :about_claimant_4, GroupClaimantSection, :group_claimants, "claims.additional_claimants.person", number: 4
+          section :about_claimant_5, GroupClaimantSection, :group_claimants, "claims.additional_claimants.person", number: 5
+          section :about_claimant_6, GroupClaimantSection, :group_claimants, "claims.additional_claimants.person", number: 6
+
           element :add_more_claimants_link, :submit_text, 'claims.additional_claimants.add_fields'
           element :save_and_continue_button, :submit_text, 'helpers.submit.update', exact: false
         end
@@ -184,9 +123,9 @@ module EtFullSystem
           group_claims_csv = data[0].dig(:group_claims_csv)
           if data.length > 1 || group_claims_csv.present?
             main_content.additional_claimants_of_collection_type.yes.click
-            data.each.with_index do |claim, idx|
+            data.each.with_index do |claimant, idx|
               add_more_claimants unless idx == 0
-              populate_group_claim_section(data[0].to_h, idx + 2)
+              populate_group_claim_section(claimant.to_h, idx + 2)
             end
             main_content.number_claimants_info.csv_upload_link.click if group_claims_csv.present?
           else
